StorageHelperAIOrchestraService/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
